#323 update most deprecated :  - getPrefetch long -> int  - remove WaitStrategy.throwAlert  - remove Exceptions.CANCEL_STACKTRACE  - remove xxx(boolean delayError) from public scope (delayError suffix)  - remove Operators.addAndGet(AtomicLong)  - remove FluxSink#serialize  - remove  Flux, buffer(int, long) -> bufferTimeout, window(int, long), delay(Duration), zip(Iterable)

+++ src/main/java/reactor/core/publisher/FluxAutoConnect.java
@@ -64,7 +64,7 @@ final class FluxAutoConnect<T> extends Flux<T>
 	}
 
 	@Override
-	public long getPrefetch() {
+	public int getPrefetch() {
 		return source.getPrefetch();
 	}
