+++ src/main/java/reactor/core/publisher/ParallelFluxHide.java
@@ -34,7 +34,7 @@ final class ParallelFluxHide<T> extends ParallelFlux<T> implements Scannable{
 	}
 
 	@Override
-	public long getPrefetch() {
+	public int getPrefetch() {
 		return source.getPrefetch();
 	}
